refactor: remove deprecated user controller and route files, enhancing clarity and maintainability in authentication management
