--- backend/routes/auth/controllers/index.js
@@ -1,13 +1,0 @@
-// Export auth controllers
-export * from './auth/login.js';
-export * from './auth/logout.js';
-export * from './auth/signup.js';
-export * from './auth/refresh.js';
-// export * from './auth/resetPassword.js';
-// export * from './auth/verify.js';
-
-// Export user controllers
-// export * from './user/getUser.js';
-// export * from './user/getAllUsers.js';
-// export * from './user/updateUser.js';
-// export * from './user/deleteUser.js';
