--- backend/routes/auth/controllers/user/index.js
@@ -1,11 +1,0 @@
-export * from './getUser.js';
-export * from './getAllUsers.js';
-export * from './updateUser.js';
-export * from './deleteUser.js';
-
-export default {
-  getAllUsers,
-  getUserById,
-  updateUser,
-  deleteUser
-};
