+++ backend/routes/user/delete-user/controller.js
@@ -1,4 +1,4 @@
-import User from '../../../../models/User.js';
+import User from '../../../models/User.js';
 
 /**
  * Delete user
