+++ backend/routes/user/get-user/controller.js
@@ -1,4 +1,4 @@
-import User from '../../../../models/User.js';
+import User from '../../../models/User.js';
 
 export const getUserById = async (req, res) => {
   try {
